move funcs into kafka_msg_handler
