--- koku/masu/external/downloader/ocp/__init__.py
@@ -1,1 +1,0 @@
-# noqa
