--- koku/masu/test/external/downloader/ocp/__init__.py
@@ -1,1 +1,0 @@
-# noqa
